feat: added pre-existing main notebook & PyTorch neural network template
Added Navigation.ipynb notebook which was originally meant to be a main for the project. I will use it mainly to observe agent behavior in the environment.
pytorch_base_network.py has been copied over from my pyTorch learning project and it eases the creation of simple linear networks with FC layers and RELUs