feat: base network forward on numpy
a new forward_np method has been added to the base neural networks. This allows passing numpy arrays instead of tensors for forward operation. As np arrays are not linked to gradient computation, methods on np arrays like this can only be used for exploitation, not for training the network.