feat: zero_grad and dones
My implementation was not zeroing gradients, the gradients would accumulate in subsequent .backward() calls. 
Additionally i was including in the target predicted future cumulative rewards even when the end of the episode was observed. (In the Banana environment this should not be a problem as the episode does not end in a specific state.)
The  current version and parametrization makes sense o cart pole but does not solve yet