feat: fixed pytorch_base_network dropout and improved its test
improvements in test_pytorch_base_network code revelaed that the neural networks would not give same results in subsequent "forward" calls in evaluation mode.
This was due to the fact that dropout was implemented as a simple forward call like Relu, it was thus not visible to the model and would not be set correctly in eval mode. This was corrected.
Test on load and learning for pytorch base network were based on visual inspection of value, they are now been implemented via assert.
parameter save structure has been slightly changed, now network parameters have been saved as a dictionary, which is converted to/from named tuple in code